MockSyncStrategy: Implement request decoding
Until now, MockSyncStrategy did not save requests in the connection
object. This breaks the automatic Reader generation in
connection._get_entries, rendering connection.entries unusable.

This patch removes some seemingly redundant logic that separating the
message type and request from the ldap message in the
BaseStrategy.decode_request staticmethod. Both of these values are
already known by the callers, allowing it to be used in the
MockSyncStrategy.