strategy/reusable: correctly refresh server info
It turns out that Connection._fire_deferred() is a no-op for non-lazy
connections. Because of this the pooled connections of ReusableStrategy
do not correctly get their schema refreshed when
Connection.refresh_server_info() is called on a connection with reusable
strategy.

This causes performing operations without the schema being present if we're
using ServerPool() with multiple servers and leads to unexpected types
returned by Connection.search().

References: jira PAM-7321