core/pooling: use a namedtuple to store server state
Instead of constructing a list of three items to store the state of a
server in a pool, this change introduces the ServerState namedtuple.

This has the advantage that parts of the tuple can be referenced by name,
so instead of writing servers[offset][0] we can now write
server_states[offset].server.