Added support for specifying a source address and source port or source port list when opening an LDAP connection.
In many simple environments, just opening a connection to a remote server at some destination address
and port is all you need.
However, sometimes you're working an environment with multiple network partitions which aren't mutually
routable. Or you're dealing with a firewall that only allows connections through if they originate from
a specific source address or address range, and you're on a machine with multiple identities from a
networking perspective.
Alternatively, you may have multiple network identities, and you're doing mutual TLS so you want to make
sure you pick the correct one. Or you just want to control source network identity at the application
level, rather than having it be controlled via ip tables or network switch configurations. Maybe an
application has a designated range of local ports that it owns, and you want to confine it to that
range within the application rather than by using a container.

Regardless of the specific reason, there's a number of use cases for specifying a source ip/hostname to
try to use when binding the socket for a connection, as well as for specifying a source port/port list
for binding the socket locally.
This change adds support for specifying such a source address, a source port, or source port list.

The idea with this change is that source address/port are attributes of the connection, rather than the
server definition. However, to faciliate checking server availability (which is done via the server
object) with this functionality, a source address, source port, and source port list can be specified
when checking server availability as well via optional arguments.