Add channel binding support for SASL + TLS
When building the connection base on SASL(Kerberos) + TLS to connect a Windows
ldap server (aka Domain Controller), we have to enforce ldap channel binding for
the connection otherwise the server will reject the connection.
See: https://support.microsoft.com/en-us/help/4520412/2020-ldap-channel-binding-and-ldap-signing-requirements-for-windows

This change also consider the useage of the SASL without TLS connection, so
won't calculate the channel binding token for that case.

Signed-off-by: Yang Qian <yang.qian@citrix.com>