Added configuration for testing against Windows AD 2019 running on an Azure instance.
Added user credentials, ous, etc. for an AD 2019 configuration that's accessible from
the public internet.
Added some extra detail for failed user_add calls, to make debugging any failures
easier, as in some cases the only error printed was the downstream error encountered
when deleting or modifying the user failed.

Also updated tests searching with a -2 filter to use -3 because -2019 is in the domain
name and caused problems.