Fix issues with unicode strings in python 2.
Unicode referrals aren't handled properly in python 2
because python 2 can't convert them to strings.
A few other attributes may also suffer from this.

This change adds a helper that attempts string conversion,
and does a backslash replace method of encoding unicode
to ascii if basic conversion fails. So for python 3
the behavior is unaffected, while for python 2 this will
hopefully avoid errors.